Refactoring pour mieux séparer le code
