+++ src/modules/chaleur-renouvelable/client/hooks/useAddressEligibility.ts
@@ -2,25 +2,11 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 
 import { searchBANAddresses } from '@/modules/ban/client';
 import type { BANAddressFeature } from '@/modules/ban/types';
-import type { BatEnrBatiment, BatEnrBatimentsSelectionContext } from '@/modules/chaleur-renouvelable/constants';
+import { type BatEnrInfo, EMPTY_BAT_ENR_INFO, getBatEnrInfoFromBatiment } from '@/modules/chaleur-renouvelable/bat-enr';
+import type { BatEnrBatiment } from '@/modules/chaleur-renouvelable/constants';
 import { toastErrors } from '@/modules/notification';
 import trpc from '@/modules/trpc/client';
-import type { HeatNetworksResponse } from '@/types/HeatNetworksResponse';
-
-type BatEnrInfo = {
-  architecturalProtectionAc1: boolean;
-  architecturalProtectionAc2: boolean;
-  architecturalProtectionAc3: boolean;
-  architecturalProtectionAc4: boolean;
-  architecturalProtectionAc4bis: boolean;
-  geothermiePossible: boolean;
-  geothermalNappePotential: number | null;
-  geothermalNappeGmi: number | null;
-  geothermalSondeGmi: number | null;
-  hasGeothermalProbeSpace: boolean | null;
-  planProtectionAtmosphere: boolean;
-  solarThermalCoverage: number | null;
-};
+import type { HeatNetwork } from '@/types/HeatNetworksResponse';
 
 type EligibilityState = {
   geoAddress?: BANAddressFeature;
@@ -29,34 +15,12 @@ type EligibilityState = {
   selectedBatEnrBatiment?: BatEnrBatiment;
   codeDepartement: string;
   temperatureRef: number | null;
-  eligibiliteReseauChaleur: HeatNetworksResponse | null;
+  eligibiliteReseauChaleur: HeatNetwork | null;
   shouldSelectBatEnrBatiment: boolean;
 };
-
-type BatEnrLookupResult = {
-  batEnr: BatEnrInfo;
-  batEnrBatiments: BatEnrBatiment[];
-  selectedBatEnrBatiment?: BatEnrBatiment;
-  shouldSelectBatEnrBatiment: boolean;
-};
-
-type TrpcUtils = ReturnType<typeof trpc.useUtils>;
 
 const emptyState: EligibilityState = {
-  batEnr: {
-    architecturalProtectionAc1: false,
-    architecturalProtectionAc2: false,
-    architecturalProtectionAc3: false,
-    architecturalProtectionAc4: false,
-    architecturalProtectionAc4bis: false,
-    geothermalNappeGmi: null,
-    geothermalNappePotential: null,
-    geothermalSondeGmi: null,
-    geothermiePossible: false,
-    hasGeothermalProbeSpace: null,
-    planProtectionAtmosphere: false,
-    solarThermalCoverage: null,
-  },
+  batEnr: EMPTY_BAT_ENR_INFO,
   batEnrBatiments: [],
   codeDepartement: '',
   eligibiliteReseauChaleur: null,
@@ -64,93 +28,6 @@ const emptyState: EligibilityState = {
   selectedBatEnrBatiment: undefined,
   shouldSelectBatEnrBatiment: false,
   temperatureRef: null,
-};
-
-const getBatEnrInfoFromBatiment = (batEnrDetails?: BatEnrBatiment | null): BatEnrInfo => {
-  return {
-    architecturalProtectionAc1: batEnrDetails?.ac1 ?? false,
-    architecturalProtectionAc2: batEnrDetails?.ac2 ?? false,
-    architecturalProtectionAc3: batEnrDetails?.ac3 ?? false,
-    architecturalProtectionAc4: batEnrDetails?.ac4 ?? false,
-    architecturalProtectionAc4bis: batEnrDetails?.ac4bis ?? false,
-    geothermalNappeGmi: batEnrDetails?.gmi_nappe_200 != null ? Number(batEnrDetails.gmi_nappe_200) : null,
-    geothermalNappePotential: batEnrDetails?.pot_nappe != null ? Number(batEnrDetails.pot_nappe) : null,
-    geothermalSondeGmi: batEnrDetails?.gmi_sonde_200 != null ? Number(batEnrDetails.gmi_sonde_200) : null,
-    geothermiePossible: [1, 2].includes(Number(batEnrDetails?.gmi_nappe_200)) || [1, 2].includes(Number(batEnrDetails?.gmi_sonde_200)),
-    hasGeothermalProbeSpace: batEnrDetails?.place_nappe ?? null,
-    planProtectionAtmosphere: batEnrDetails?.etat_ppa === 'PPA Validés',
-    solarThermalCoverage: batEnrDetails?.couv_st_ecs_2025 != null ? Number(batEnrDetails.couv_st_ecs_2025) : null,
-  };
-};
-
-const getBatEnrLookupResult = async ({
-  geoAddress,
-  selectedBatimentConstructionId,
-  trpcUtils,
-}: {
-  geoAddress: BANAddressFeature;
-  selectedBatimentConstructionId?: string | null;
-  trpcUtils: TrpcUtils;
-}): Promise<BatEnrLookupResult> => {
-  const [lon, lat] = geoAddress.geometry.coordinates;
-  const banId = geoAddress.properties.id;
-  const selectionContext = await trpcUtils.client.batEnr.getBatEnrBatimentsSelectionContextByBanId
-    .query({ banId })
-    .catch((): BatEnrBatimentsSelectionContext => ({ batiments: [], preselectedBatimentConstructionId: null }));
-  const batEnrBatiments = selectionContext.batiments;
-
-  if (selectedBatimentConstructionId) {
-    const selectedBatEnrBatiment = batEnrBatiments.find((batiment) => batiment.batiment_construction_id === selectedBatimentConstructionId);
-
-    if (selectedBatEnrBatiment) {
-      return {
-        batEnr: getBatEnrInfoFromBatiment(selectedBatEnrBatiment),
-        batEnrBatiments,
-        selectedBatEnrBatiment,
-        shouldSelectBatEnrBatiment: false,
-      };
-    }
-  }
-
-  const preselectedBatEnrBatiment = selectionContext.preselectedBatimentConstructionId
-    ? batEnrBatiments.find((batiment) => batiment.batiment_construction_id === selectionContext.preselectedBatimentConstructionId)
-    : undefined;
-
-  if (preselectedBatEnrBatiment) {
-    return {
-      batEnr: getBatEnrInfoFromBatiment(preselectedBatEnrBatiment),
-      batEnrBatiments,
-      selectedBatEnrBatiment: preselectedBatEnrBatiment,
-      shouldSelectBatEnrBatiment: batEnrBatiments.length > 1,
-    };
-  }
-
-  if (batEnrBatiments.length === 1) {
-    return {
-      batEnr: getBatEnrInfoFromBatiment(batEnrBatiments[0]),
-      batEnrBatiments,
-      selectedBatEnrBatiment: batEnrBatiments[0],
-      shouldSelectBatEnrBatiment: false,
-    };
-  }
-
-  if (batEnrBatiments.length > 1) {
-    return {
-      batEnr: getBatEnrInfoFromBatiment(null),
-      batEnrBatiments,
-      selectedBatEnrBatiment: undefined,
-      shouldSelectBatEnrBatiment: true,
-    };
-  }
-
-  const batEnrDetails = await trpcUtils.client.batEnr.getBatEnrBatimentDetails.query({ lat, lon }).catch(() => null);
-
-  return {
-    batEnr: getBatEnrInfoFromBatiment(batEnrDetails),
-    batEnrBatiments: batEnrDetails ? [batEnrDetails] : [],
-    selectedBatEnrBatiment: batEnrDetails ?? undefined,
-    shouldSelectBatEnrBatiment: false,
-  };
 };
 
 export function useAddressEligibility(adresse: string | null, selectedBatimentConstructionId?: string | null) {
@@ -175,27 +52,24 @@ export function useAddressEligibility(adresse: string | null, selectedBatimentCo
         const [lon, lat] = geoAddress.geometry.coordinates;
         const { city, citycode } = geoAddress.properties;
 
-        const [batEnrLookup, infos, eligibiliteReseauChaleur] = await Promise.all([
-          getBatEnrLookupResult({ geoAddress, selectedBatimentConstructionId: selectedBatimentConstructionIdRef.current, trpcUtils }),
-          trpcUtils.client.batEnr.getLocationInfos.query({
-            city,
-            cityCode: citycode,
-          }),
-          trpcUtils.client.reseaux.eligibilityStatus.query({
-            lat,
-            lon,
-          }),
-        ]);
+        const addressEligibilityContext = await trpcUtils.client.batEnr.getAddressEligibilityContext.query({
+          banId: geoAddress.properties.id,
+          city,
+          cityCode: citycode,
+          lat,
+          lon,
+          selectedBatimentConstructionId: selectedBatimentConstructionIdRef.current,
+        });
 
         setState({
-          batEnr: batEnrLookup.batEnr,
-          batEnrBatiments: batEnrLookup.batEnrBatiments,
-          codeDepartement: infos?.departement_id ?? '',
-          eligibiliteReseauChaleur,
+          batEnr: addressEligibilityContext.batEnr,
+          batEnrBatiments: addressEligibilityContext.batEnrBatiments,
+          codeDepartement: addressEligibilityContext.codeDepartement,
+          eligibiliteReseauChaleur: addressEligibilityContext.eligibiliteReseauChaleur,
           geoAddress,
-          selectedBatEnrBatiment: batEnrLookup.selectedBatEnrBatiment,
-          shouldSelectBatEnrBatiment: batEnrLookup.shouldSelectBatEnrBatiment,
-          temperatureRef: infos?.temperature_ref_altitude_moyenne != null ? Number(infos.temperature_ref_altitude_moyenne) : null,
+          selectedBatEnrBatiment: addressEligibilityContext.selectedBatEnrBatiment,
+          shouldSelectBatEnrBatiment: addressEligibilityContext.shouldSelectBatEnrBatiment,
+          temperatureRef: addressEligibilityContext.temperatureRef,
         });
       } finally {
         setIsEligibilityLoading(false);
