+++ src/modules/chaleur-renouvelable/client/modesChauffageData.ts
@@ -1,0 +1,13 @@
+export type {
+  IncompatibleSolutionRow,
+  ModeDeChauffage,
+  ModeDeChauffageEnriched,
+  ModeDeChauffageId,
+  ModeDeChauffageUsage,
+  PrerequisiteRow,
+  PrerequisiteStatus,
+  Situation,
+} from '@/modules/chaleur-renouvelable/constants';
+
+export { modesDeChauffage } from './heating-modes/catalog';
+export { getIncompatibleSolutionRows, getModesDeChauffage, improveDpe } from './heating-modes/selectors';
